# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Implementation Api CV",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/telefono": {
            "get": {
                "tags": [
                    "Teléfonos"
                ],
                "description": "Get phone list",
                "operationId": "4ffd506a53cadd4b3b56031da131abb1",
                "responses": {
                    "200": {
                        "description": "Listado de Teléfonos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Teléfonos"
                ],
                "description": "Store a newly created resource in database",
                "operationId": "5d4f6d77684404f9d4b8983318245f8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "description": "Número de Teléfono",
                                        "type": "integer"
                                    },
                                    "tipo_telefono_id": {
                                        "description": "Identificación del tipo de teléfono",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "Identificación del Usuario",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crear un Teléfono"
                    }
                }
            }
        },
        "/api/telefono/{id}": {
            "get": {
                "tags": [
                    "Teléfonos"
                ],
                "description": "Get a record of phone type",
                "operationId": "21526c365592c39ad4160714a62da507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id for Phone",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener un tipo de teléfono"
                    }
                }
            },
            "put": {
                "tags": [
                    "Teléfonos"
                ],
                "description": "Update a Phone",
                "operationId": "d78238bb53fad42316e4395e7414b666",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "description": "number",
                                        "type": "integer"
                                    },
                                    "tipo_telefono_id": {
                                        "description": "Phone type identification",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "User Identification",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actualizar un Teléfono"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Teléfonos"
                ],
                "description": "Delete a single phone on the ID supplied",
                "operationId": "29172e39dd07e14e4855ce10ac523275",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Phone to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminar un telefono"
                    }
                }
            }
        },
        "/api/tipo-telefono": {
            "get": {
                "tags": [
                    "Tipos de Teléfonos"
                ],
                "description": "Get all the types",
                "operationId": "952af30ae3459c942909012bd2228640",
                "responses": {
                    "200": {
                        "description": "Listado de tipo de Teléfono"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de Teléfonos"
                ],
                "description": "Store a newly created resource in database",
                "operationId": "b645c23c311fc3fdd38e5df1eb9f5e1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "Descripción del Tipo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crear un tipo"
                    }
                }
            }
        },
        "/api/tipo-telefono/{tipo}": {
            "get": {
                "tags": [
                    "Tipos de Teléfonos"
                ],
                "description": "Get a record of phone type",
                "operationId": "90941f672ecf12124fbf9467a1d8eba3",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "path",
                        "description": "id for type",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener un tipo de teléfono"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Teléfonos"
                ],
                "description": "Update a type",
                "operationId": "237c085cb9204ff53384bb022c228e44",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actualizar un tipo"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos de Teléfonos"
                ],
                "description": "Delete a single type on the ID supplied",
                "operationId": "f5786d5ff5219db1a059e576e8cac5c8",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "path",
                        "description": "ID of type to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminar un tipo de telefono"
                    }
                }
            }
        },
...
    }
}